requirements.txt
pandas
scikit-learn
opencv-python
Pillow
mtcnn
keras
